Added Metro stops NearestMetroStop method
