+++ .gitignore
@@ -1,0 +1,1 @@
+backend/aspBuild/new-york-latest.osm.pbf
