+++ backend/aspBuild/csFiles/CreateDatabase.cs
@@ -99,7 +99,7 @@ class CreateDatabase
                     // populates the Node object and adds it to the database
                     MapNode tempNode = new MapNode(node.Id.Value, node.Longitude.Value, node.Latitude.Value);
                     string tempTaxiZone = taxiZones.PointInTaxiZone(node.Latitude.Value, node.Longitude.Value);
-                    List<string> tempMetroNearby = metroStops.PointInCircle(node.Latitude.Value, node.Longitude.Value);
+                    string tempMetroNearby = metroStops.NearestMetroStop(node.Latitude.Value, node.Longitude.Value);
                     int tempRoadRank = pedestrianData.ClosestRoadRank(node.Latitude.Value, node.Longitude.Value);
                     bool ParkTrueFalse = parks.ParkTrueFalse(node.Latitude.Value, node.Longitude.Value);
