+++ backend/aspBuild/csFiles/HaversineCalculator.cs
@@ -1,0 +1,37 @@
+using System;
+
+public class HaversineCalculator
+{
+    // Generated by ChatGPT for use in MetroStops
+    // Radius of Earth in kilometers. Use 3958.8 for miles.
+    private const double EarthRadiusKm = 6371.0;
+
+    public static double CalculateDistance(double lat1, double lon1, double lat2, double lon2)
+    {
+        // Convert latitude and longitude from degrees to radians
+        double lat1Rad = ToRadians(lat1);
+        double lon1Rad = ToRadians(lon1);
+        double lat2Rad = ToRadians(lat2);
+        double lon2Rad = ToRadians(lon2);
+
+        // Differences in coordinates
+        double dLat = lat2Rad - lat1Rad;
+        double dLon = lon2Rad - lon1Rad;
+
+        // Haversine formula
+        double a = Math.Sin(dLat / 2) * Math.Sin(dLat / 2) +
+                   Math.Cos(lat1Rad) * Math.Cos(lat2Rad) *
+                   Math.Sin(dLon / 2) * Math.Sin(dLon / 2);
+        double c = 2 * Math.Asin(Math.Sqrt(a));
+
+        // Distance in kilometers
+        double distance = EarthRadiusKm * c;
+
+        return distance;
+    }
+
+    private static double ToRadians(double degrees)
+    {
+        return degrees * (Math.PI / 180);
+    }
+}
